Reorganize aspath module
