+++ src/models/bgp/aspath/mod.rs
@@ -1,11 +1,17 @@
 use crate::models::*;
 use itertools::Itertools;
-use std::borrow::Cow;
 use std::fmt::{Display, Formatter};
 use std::hash::{Hash, Hasher};
-use std::iter::FromIterator;
-use std::marker::PhantomData;
 use std::mem::discriminant;
+
+pub mod iters;
+pub use iters::*;
+
+#[cfg(feature = "serde")]
+mod serde_impl;
+
+#[cfg(test)]
+mod tests;
 
 /// Enum of AS path segment.
 #[derive(Debug, Clone)]
@@ -46,16 +52,6 @@ impl AsPathSegment {
     /// Returns true if this segment has a length of 0.
     pub fn is_empty(&self) -> bool {
         self.as_ref().is_empty()
-    }
-
-    /// Get an iterator over the ASNs within this path segment
-    pub fn iter(&self) -> <&'_ Self as IntoIterator>::IntoIter {
-        self.into_iter()
-    }
-
-    /// Get a mutable iterator over the ASNs within this path segment
-    pub fn iter_mut(&mut self) -> <&'_ mut Self as IntoIterator>::IntoIter {
-        self.into_iter()
     }
 
     /// Gets if a segment represents the local members of an autonomous system confederation.
@@ -134,45 +130,6 @@ impl AsPathSegment {
                 }
             }
         }
-    }
-}
-
-impl IntoIterator for AsPathSegment {
-    type Item = Asn;
-    type IntoIter = std::vec::IntoIter<Asn>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        let (AsPathSegment::AsSequence(x)
-        | AsPathSegment::AsSet(x)
-        | AsPathSegment::ConfedSequence(x)
-        | AsPathSegment::ConfedSet(x)) = self;
-        x.into_iter()
-    }
-}
-
-impl<'a> IntoIterator for &'a AsPathSegment {
-    type Item = &'a Asn;
-    type IntoIter = std::slice::Iter<'a, Asn>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        let (AsPathSegment::AsSequence(x)
-        | AsPathSegment::AsSet(x)
-        | AsPathSegment::ConfedSequence(x)
-        | AsPathSegment::ConfedSet(x)) = self;
-        x.iter()
-    }
-}
-
-impl<'a> IntoIterator for &'a mut AsPathSegment {
-    type Item = &'a mut Asn;
-    type IntoIter = std::slice::IterMut<'a, Asn>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        let (AsPathSegment::AsSequence(x)
-        | AsPathSegment::AsSet(x)
-        | AsPathSegment::ConfedSequence(x)
-        | AsPathSegment::ConfedSet(x)) = self;
-        x.iter_mut()
     }
 }
 
@@ -273,95 +230,10 @@ impl PartialEq for AsPathSegment {
 
 impl Eq for AsPathSegment {}
 
-/// This is not a perfect solution since it is theoretically possible that a path could be created
-/// with more variations than a u64. That being said, the chances of such a thing occurring are
-/// essentially non-existent unless a BGP peer begins announcing maliciously constructed paths.
-struct AsPathNumberedRouteIter<'a> {
-    path: &'a [AsPathSegment],
-    index: usize,
-    route_num: u64,
-}
-
-impl<'a> Iterator for AsPathNumberedRouteIter<'a> {
-    type Item = Asn;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            match self.path.first()? {
-                AsPathSegment::AsSequence(x) => match x.get(self.index) {
-                    None => {
-                        self.index = 0;
-                        self.path = &self.path[1..];
-                    }
-                    Some(asn) => {
-                        self.index += 1;
-                        return Some(*asn);
-                    }
-                },
-                AsPathSegment::AsSet(x) => {
-                    self.path = &self.path[1..];
-                    if x.is_empty() {
-                        return Some(Asn::RESERVED);
-                    }
-
-                    let asn = x[(self.route_num % x.len() as u64) as usize];
-                    self.route_num /= x.len() as u64;
-                    return Some(asn);
-                }
-                _ => self.path = &self.path[1..],
-            }
-        }
-    }
-}
-
-pub struct AsPathRouteIter<'a, D> {
-    path: Cow<'a, [AsPathSegment]>,
-    route_num: u64,
-    total_routes: u64,
-    _phantom: PhantomData<D>,
-}
-
-impl<'a, D> Iterator for AsPathRouteIter<'a, D>
-where
-    D: FromIterator<Asn>,
-{
-    type Item = D;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.route_num >= self.total_routes {
-            return None;
-        }
-
-        // Attempt to speed up what is by far the most common case (a path of a single sequence)
-        if self.route_num == 0 && self.path.len() == 1 {
-            if let AsPathSegment::AsSequence(sequence) = &self.path[0] {
-                let route = D::from_iter(sequence.iter().copied());
-                self.route_num += 1;
-                return Some(route);
-            }
-        }
-
-        let route_asn_iter = AsPathNumberedRouteIter {
-            path: self.path.as_ref(),
-            index: 0,
-            route_num: self.route_num,
-        };
-
-        self.route_num += 1;
-        Some(D::from_iter(route_asn_iter))
-    }
-}
-
 #[derive(Debug, PartialEq, Clone, Eq, Default, Hash)]
 pub struct AsPath {
     pub segments: Vec<AsPathSegment>,
 }
-
-// Define iterator type aliases. The storage mechanism and by extension the iterator types may
-// change later, but these types should remain consistent.
-pub type SegmentIter<'a> = std::slice::Iter<'a, AsPathSegment>;
-pub type SegmentIterMut<'a> = std::slice::IterMut<'a, AsPathSegment>;
-pub type SegmentIntoIter = std::vec::IntoIter<AsPathSegment>;
 
 impl AsPath {
     pub fn new() -> AsPath {
@@ -550,31 +422,6 @@ impl AsPath {
         )
     }
 
-    pub fn iter_segments(&self) -> SegmentIter<'_> {
-        self.segments.iter()
-    }
-
-    pub fn iter_segments_mut(&mut self) -> SegmentIterMut<'_> {
-        self.segments.iter_mut()
-    }
-
-    pub fn into_segments_iter(self) -> SegmentIntoIter {
-        self.segments.into_iter()
-    }
-
-    /// Gets an iterator over all possible routes this path represents.
-    pub fn iter_routes<D>(&self) -> AsPathRouteIter<'_, D>
-    where
-        D: FromIterator<Asn>,
-    {
-        AsPathRouteIter {
-            path: Cow::Borrowed(&self.segments),
-            route_num: 0,
-            total_routes: self.num_route_variations(),
-            _phantom: PhantomData,
-        }
-    }
-
     /// Construct AsPath from AS_PATH and AS4_PATH
     ///
     /// <https://datatracker.ietf.org/doc/html/rfc6793#section-4.2.3>
@@ -628,19 +475,6 @@ impl AsPath {
         Some(AsPath { segments: new_segs })
     }
 
-    /// Iterate through the originating ASNs of this path. This functionality is provided for
-    /// completeness, but in almost all cases this iterator should only contain a single element.
-    /// Alternatively, [AsPath::get_singular_origin] can be used if
-    pub fn iter_origins(&self) -> impl '_ + Iterator<Item = Asn> {
-        let origin_slice = match self.segments.last() {
-            Some(AsPathSegment::AsSequence(v)) => v.last().map(std::slice::from_ref).unwrap_or(&[]),
-            Some(AsPathSegment::AsSet(v)) => v.as_ref(),
-            _ => &[],
-        };
-
-        origin_slice.iter().copied()
-    }
-
     /// This function serves as a alternative to [AsPath::iter_origins] which attempts to make the
     /// assumption that a path can only have exactly one origin. If a path does not have exactly 1
     /// origin (such as when empty or ending in a set), then `None` will be returned instead.
@@ -649,31 +483,6 @@ impl AsPath {
             Some(AsPathSegment::AsSequence(v)) => v.last().copied(),
             Some(AsPathSegment::AsSet(v)) if v.len() == 1 => Some(v[0]),
             _ => None,
-        }
-    }
-}
-
-/// Iterates over all route variations the given `AsPath` represents.
-impl<'a> IntoIterator for &'a AsPath {
-    type Item = Vec<Asn>;
-    type IntoIter = AsPathRouteIter<'a, Vec<Asn>>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.iter_routes()
-    }
-}
-
-/// Iterates over all route variations the given `AsPath` represents.
-impl IntoIterator for AsPath {
-    type Item = Vec<Asn>;
-    type IntoIter = AsPathRouteIter<'static, Vec<Asn>>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        AsPathRouteIter {
-            total_routes: self.num_route_variations(),
-            path: Cow::Owned(self.segments),
-            route_num: 0,
-            _phantom: PhantomData,
         }
     }
 }
@@ -712,303 +521,5 @@ impl Display for AsPath {
         }
 
         Ok(())
-    }
-}
-
-#[cfg(feature = "serde")]
-mod serde_impl {
-    use super::*;
-    use serde::de::{SeqAccess, Visitor};
-    use serde::ser::SerializeSeq;
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-    use std::borrow::Cow;
-
-    /// Segment type names using names from RFC3065.
-    ///
-    /// <https://datatracker.ietf.org/doc/html/rfc3065#section-5>
-    #[allow(non_camel_case_types)]
-    #[derive(Serialize, Deserialize)]
-    enum SegmentType {
-        AS_SET,
-        AS_SEQUENCE,
-        AS_CONFED_SEQUENCE,
-        AS_CONFED_SET,
-    }
-
-    #[derive(Serialize, Deserialize)]
-    struct VerboseSegment<'s> {
-        ty: SegmentType,
-        values: Cow<'s, [Asn]>,
-    }
-
-    impl Serialize for AsPathSegment {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            let (ty, elements) = match self {
-                AsPathSegment::AsSequence(x) => (SegmentType::AS_SEQUENCE, x.as_ref()),
-                AsPathSegment::AsSet(x) => (SegmentType::AS_SET, x.as_ref()),
-                AsPathSegment::ConfedSequence(x) => (SegmentType::AS_CONFED_SEQUENCE, x.as_ref()),
-                AsPathSegment::ConfedSet(x) => (SegmentType::AS_CONFED_SET, x.as_ref()),
-            };
-
-            let verbose = VerboseSegment {
-                ty,
-                values: Cow::Borrowed(elements),
-            };
-
-            verbose.serialize(serializer)
-        }
-    }
-
-    impl<'de> Deserialize<'de> for AsPathSegment {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            let verbose = VerboseSegment::deserialize(deserializer)?;
-
-            let values = verbose.values.into_owned();
-            match verbose.ty {
-                SegmentType::AS_SET => Ok(AsPathSegment::AsSet(values)),
-                SegmentType::AS_SEQUENCE => Ok(AsPathSegment::AsSequence(values)),
-                SegmentType::AS_CONFED_SEQUENCE => Ok(AsPathSegment::ConfedSequence(values)),
-                SegmentType::AS_CONFED_SET => Ok(AsPathSegment::ConfedSet(values)),
-            }
-        }
-    }
-
-    /// Check if we can serialize an `AsPath` using the simplified format and get the number of
-    /// elements to do so. The ambiguities that could prevent us from doing so are confederation
-    /// segments and adjacent sequence segments.
-    fn simplified_format_len(segments: &[AsPathSegment]) -> Option<usize> {
-        let mut elements = 0;
-        let mut prev_was_sequence = false;
-        for segment in segments {
-            match segment {
-                AsPathSegment::AsSequence(seq) if !prev_was_sequence => {
-                    prev_was_sequence = true;
-                    elements += seq.len();
-                }
-                AsPathSegment::AsSet(_) => {
-                    prev_was_sequence = false;
-                    elements += 1;
-                }
-                _ => return None,
-            }
-        }
-
-        Some(elements)
-    }
-
-    /// # Serialization format
-    /// For the sake of readability and ease of use within other applications, there are verbose and
-    /// simplified variants for serialization.
-    ///
-    /// ## Simplified format
-    /// The simplified format is the default preferred serialization format. This format does not
-    /// cover confederation segments and involves a single list of ASNs within the path sequence.
-    /// For sets, a list of set members is used in place of an ASN.
-    /// ```rust
-    /// # use bgpkit_parser::models::{Asn, AsPath};
-    /// # use bgpkit_parser::models::AsPathSegment::*;
-    ///
-    /// let a: AsPath = serde_json::from_str("[123, 942, 102]").unwrap();
-    /// let b: AsPath = serde_json::from_str("[231, 432, [643, 836], 352]").unwrap();
-    ///
-    /// assert_eq!(&a.segments, &[
-    ///     AsSequence(vec![Asn::from(123), Asn::from(942), Asn::from(102)])
-    /// ]);
-    /// assert_eq!(&b.segments, &[
-    ///     AsSequence(vec![Asn::from(231), Asn::from(432)]),
-    ///     AsSet(vec![Asn::from(643), Asn::from(836)]),
-    ///     AsSequence(vec![Asn::from(352)])
-    /// ]);
-    /// ```
-    ///
-    /// ## Verbose format
-    /// The verbose format serves as the fallback format for when the simplified format can not be
-    /// used due to ambiguity. This happens when confederation segments are present, or multiple
-    /// sequences occur back to back. In this format, segments are explicitly seperated and labeled.
-    /// Segment types, denoted by the `ty` field, correspond to the names used within RFC3065
-    /// (`AS_SET`, `AS_SEQUENCE`, `AS_CONFED_SEQUENCE`, `AS_CONFED_SET`).
-    /// ```rust
-    /// # use bgpkit_parser::models::{Asn, AsPath};
-    /// # use bgpkit_parser::models::AsPathSegment::*;
-    ///
-    /// let a = r#"[
-    ///     { "ty": "AS_CONFED_SEQUENCE", "values": [123, 942] },
-    ///     { "ty": "AS_SEQUENCE", "values": [773] },
-    ///     { "ty": "AS_SEQUENCE", "values": [382, 293] }
-    /// ]"#;
-    ///
-    /// let parsed: AsPath = serde_json::from_str(a).unwrap();
-    /// assert_eq!(&parsed.segments, &[
-    ///     ConfedSequence(vec![Asn::from(123), Asn::from(942)]),
-    ///     AsSequence(vec![Asn::from(773)]),
-    ///     AsSequence(vec![Asn::from(382), Asn::from(293)])
-    /// ]);
-    /// ```
-    impl Serialize for AsPath {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            if let Some(num_elements) = simplified_format_len(&self.segments) {
-                // Serialize simplified format
-                let mut seq_serializer = serializer.serialize_seq(Some(num_elements))?;
-
-                for segment in &self.segments {
-                    match segment {
-                        AsPathSegment::AsSequence(elements) => {
-                            elements
-                                .iter()
-                                .try_for_each(|x| seq_serializer.serialize_element(x))?;
-                        }
-                        AsPathSegment::AsSet(x) => seq_serializer.serialize_element(x)?,
-                        _ => unreachable!("simplified_format_len checked for confed segments"),
-                    }
-                }
-                return seq_serializer.end();
-            }
-
-            // Serialize verbose format
-            serializer.collect_seq(&self.segments)
-        }
-    }
-
-    struct AsPathVisitor;
-
-    impl<'de> Visitor<'de> for AsPathVisitor {
-        type Value = AsPath;
-
-        fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
-            formatter.write_str("list of AS_PATH segments")
-        }
-
-        fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-        where
-            A: SeqAccess<'de>,
-        {
-            // Technically, we can handle an input that mixes the simplified and verbose formats,
-            // but we do not want to document this behavior as it may change in future updates.
-            #[derive(Deserialize)]
-            #[serde(untagged)]
-            enum PathElement {
-                SequenceElement(Asn),
-                Set(Vec<Asn>),
-                Verbose(AsPathSegment),
-            }
-
-            let mut append_new_sequence = false;
-            let mut segments = Vec::new();
-            while let Some(element) = seq.next_element()? {
-                match element {
-                    PathElement::SequenceElement(x) => {
-                        if append_new_sequence {
-                            // If the input is mixed between verbose and regular segments, this flag
-                            // is used to prevent appending to a verbose sequence.
-                            append_new_sequence = false;
-                            segments.push(AsPathSegment::AsSequence(Vec::new()));
-                        }
-
-                        if let Some(AsPathSegment::AsSequence(last_sequence)) = segments.last_mut()
-                        {
-                            last_sequence.push(x);
-                        } else {
-                            segments.push(AsPathSegment::AsSequence(vec![x]));
-                        }
-                    }
-                    PathElement::Set(values) => {
-                        segments.push(AsPathSegment::AsSet(values));
-                    }
-                    PathElement::Verbose(verbose) => {
-                        segments.push(verbose);
-                    }
-                }
-            }
-
-            Ok(AsPath { segments })
-        }
-    }
-
-    impl<'de> Deserialize<'de> for AsPath {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            deserializer.deserialize_seq(AsPathVisitor)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::models::*;
-    use itertools::Itertools;
-    use std::collections::HashSet;
-
-    #[test]
-    fn test_aspath_as4path_merge() {
-        let aspath = AsPath::from_sequence([1, 2, 3, 5]);
-        let as4path = AsPath::from_sequence([2, 3, 7]);
-        let newpath = AsPath::merge_aspath_as4path(&aspath, &as4path).unwrap();
-        assert_eq!(newpath.segments[0], AsPathSegment::sequence([1, 2, 3, 7]));
-    }
-
-    #[test]
-    fn test_get_origin() {
-        let aspath = AsPath::from_sequence([1, 2, 3, 5]);
-        let origins = aspath.get_singular_origin();
-        assert_eq!(origins.unwrap(), Asn::from(5));
-
-        let aspath = AsPath::from_segments(vec![
-            AsPathSegment::sequence([1, 2, 3, 5]),
-            AsPathSegment::set([7, 8]),
-        ]);
-        let origins = aspath.iter_origins().map_into::<u32>().collect::<Vec<_>>();
-        assert_eq!(origins, vec![7, 8]);
-    }
-
-    #[test]
-    fn test_aspath_route_iter() {
-        let path = AsPath::from_segments(vec![
-            AsPathSegment::set([3, 4]),
-            AsPathSegment::set([5, 6]),
-            AsPathSegment::sequence([7, 8]),
-        ]);
-        assert_eq!(path.route_len(), 4);
-
-        let mut routes = HashSet::new();
-        for route in &path {
-            assert!(routes.insert(route));
-        }
-
-        assert_eq!(routes.len(), 4);
-        assert!(routes.contains(&vec![
-            Asn::from(3),
-            Asn::from(5),
-            Asn::from(7),
-            Asn::from(8)
-        ]));
-        assert!(routes.contains(&vec![
-            Asn::from(3),
-            Asn::from(6),
-            Asn::from(7),
-            Asn::from(8)
-        ]));
-        assert!(routes.contains(&vec![
-            Asn::from(4),
-            Asn::from(5),
-            Asn::from(7),
-            Asn::from(8)
-        ]));
-        assert!(routes.contains(&vec![
-            Asn::from(4),
-            Asn::from(6),
-            Asn::from(7),
-            Asn::from(8)
-        ]));
     }
 }
