+++ src/models/bgp/mod.rs
@@ -6,6 +6,7 @@ pub mod community;
 pub mod elem;
 pub mod error;
 pub mod role;
+pub mod aspath;
 
 pub use attributes::*;
 pub use capabilities::*;
@@ -13,6 +14,7 @@ pub use community::*;
 pub use elem::*;
 pub use error::*;
 pub use role::*;
+pub use aspath::*;
 
 use crate::models::network::*;
 use capabilities::BgpCapabilityType;
